user_srv/requirements.txt
peewee
pymysql
passlib  加密库
grpcio
grpcio-tools [protobuf]
loguru